bug fix
git-svn-id: https://ipesvn.fzi.de/svn-open/icep/semantic-epa-parent@1504 aa18e314-994e-4353-9a36-64d395711e41

+++ semantic-epa-commons/src/main/java/de/fzi/cep/sepa/commons/config/ClientConfiguration.java
@@ -371,7 +371,7 @@ public enum ClientConfiguration {
 		return waitForMs;
 	}
 	
-	public long getKafkaBatchSize() {
+	public int getKafkaBatchSize() {
 		return kafkaBatchSize;
 	}
